requirement.txt
snowflake-connector-python
pyspark==3.4.0
ijson==3.1.4
pandas==1.3.3
pyodbc
SQLAlchemy==1.4.25
datetime==4.3
